requirements.txt
# Core Computation (GPU)
torch>=2.0.0,--index-url https://download.pytorch.org/whl/cu118
torchvision
torchaudio

# Audio Processing
numpy
librosa>=0.10.0
sounddevice
pysoundfile
cffi

# Visualization & Dashboard
visdom
requests
Pillow
websocket-client

# Utilities
scipy
tqdm
